Transformation: move pipeline executor and models to pipeline-execution package
