+++ transformation/src/pipeline-execution/jobResult.ts
@@ -1,5 +1,5 @@
 import Stats from './stats'
-import JobError from '@/pipeline-execution/sandbox/jobError'
+import JobError from './sandbox/jobError'
 
 export default interface JobResult {
   data?: object;
