+++ transformation/src/pipeline-execution/stats.ts
